face_recognition_service/requirements.txt
fastapi
uvicorn[standard]
face_recognition
opencv-python
numpy
python-multipart
pillow
